Safari-611.1.10/LayoutTests/fast/xmlhttprequest/xmlhttprequest-responsetype-json-invalid-expected.txt

Tests XMLHttpRequest.responseType of "json" with an invalid payload.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.response is null
PASS successfullyParsed is true

TEST COMPLETE
